Create README.md
Use AI technology to save the planet and live in a state of pure free expression of art, music, science and inner exploration....for peace. With each other and with ourselves...$ is nothing if we destroy all the things we would want to buy with the money we made destroying it all in the first place. As we watch others destroy our world with the belief that with enough of those $s we make from it, we can save it...eventually. Well, it's almost too late. The true CREATIVE soul left to it's freedom and a burst of peaceful and restructive technology can possibly save us all.